reduce the size of the test cases to automated testing runs faster

+++ covid_xprize/examples/predictors/conditional_lstm/tests/test_conditional_xprize_predictor.py
@@ -1,6 +1,7 @@
 # Copyright 2020 (c) Cognizant Digital Business, Evolutionary AI. All rights reserved. Issued under the Apache 2.0 License.
 
 import os
+from pathlib import Path
 import unittest
 import urllib.request
 
@@ -9,11 +10,11 @@ import pandas as pd
 from covid_xprize.examples.predictors.conditional_lstm.conditional_xprize_predictor import ConditionalXPrizePredictor
 from covid_xprize.oxford_data import load_oxford_data_trimmed
 
-ROOT_DIR = os.path.dirname(os.path.abspath(__file__))
-FIXTURES_PATH = os.path.join(ROOT_DIR, 'fixtures')
-EXAMPLE_INPUT_FILE = os.path.join(ROOT_DIR, "../../../../validation/data/2020-09-30_historical_ip.csv")
-DATA_FILE = os.path.join(FIXTURES_PATH, "OxCGRT_trimmed.csv")
-PREDICTOR_WEIGHTS = os.path.join(FIXTURES_PATH, "trained_model_weights_for_tests.h5")
+ROOT_DIR = Path(__file__).parent
+FIXTURES_PATH = ROOT_DIR / 'fixtures'
+EXAMPLE_INPUT_FILE = (ROOT_DIR / "../../../../validation/data/2020-09-30_historical_ip.csv").absolute()
+DATA_FILE = FIXTURES_PATH / "OxCGRT_trimmed.csv"
+PREDICTOR_WEIGHTS = FIXTURES_PATH / "trained_model_weights_for_tests.h5"
 
 TRAINING_START_DATE = "2020-06-01"
 TRAINING_END_DATE = "2020-07-31"
@@ -25,12 +26,16 @@ class TestConditionalXPrizePredictor(unittest.TestCase):
 
     @classmethod
     def setUpClass(cls):
-        df = load_oxford_data_trimmed(end_date=TRAINING_END_DATE, start_date=TRAINING_START_DATE)
-        df.to_csv(DATA_FILE, index=False)
+        FIXTURES_PATH.mkdir(exist_ok=True)
+        if not DATA_FILE.exists():
+            df = load_oxford_data_trimmed(end_date=TRAINING_END_DATE, start_date=TRAINING_START_DATE)
+            df.to_csv(DATA_FILE, index=False)
 
     def test_train_and_predict(self):
         predictor = ConditionalXPrizePredictor(None, DATA_FILE)
-        model = predictor.train()
+
+        # Testing on a small number of epochs, trials and geos to make sure everything is wired correctly
+        model = predictor.train(nb_epochs=2, nb_trials=2, nb_testing_geos=2, nb_training_geos=2)
         model.save_weights(PREDICTOR_WEIGHTS)
         self.assertIsNotNone(model)
